refactor(api): rename types module to api_types and update type annotations
- Rename 'types.py' to 'api_types.py' to better reflect its contents
- Update import statements in affected modules
- Replace 'Optional[str]' with 'str | None' for improved readability
